BatteryStatsProxy: Improve performance on large arrays
Aviods extensive copying and garbage collection. When created with a default constuctor,
ArrayList allocates memory for 5 elements. Once you 6th is added, new Array for 10
elements is allocated, 5 old elements are copied and then 6th is added. Same will
happen for every 5*i + 1 element. To avoid that, cache the size and use it to adjust
array size at the beginning.

	modified:   src/com/asksven/android/common/privateapiproxies/BatteryStatsProxy.java